Keep renderer observation GPU code target-scoped
